remove delete.go
